PR 2.1: cell collaborator migration
Move the five cell-geometry method bodies (get_cell_coords,
get_cell_polygons, get_cell_points, map_to_array_coordinates,
array_to_map_coordinates) from src/pyramids/dataset/ops/cell.py:Cell
into src/pyramids/dataset/_collaborators.py:Cell. Inside the
collaborator, every read of Dataset state (geotransform, no_data_value,
read_array, top_left_corner, cell_size, x, y, _get_epsg, logger) now
goes through self._ds.<state> via the weakref proxy.

src/pyramids/dataset/dataset.py:

* Drop _CellMixin from the import alias list and from Dataset's MRO.
* Add five Dataset facade methods that thin-forward to self.cell.<X>
  so the public ds.get_cell_coords(...) API is preserved.

src/pyramids/dataset/ops/__init__.py: drop Cell import and __all__
entry. src/pyramids/dataset/ops/cell.py: deleted (no other content).
docs/reference/dataset/cell.md: redirect mkdocstrings reference at the
new module path pyramids.dataset._collaborators.Cell.

tests/dataset/test_collaborators.py: split FORWARDING_METHODS into two
parametrised lists. The 5 cell rows move to a new FACADE_METHODS list,
covered by a new TestFacadeDelegation test class that verifies the
direction inversion (Dataset facade calls collaborator method, not the
other way around). The 40 remaining Stage-1 forwarder rows still go
through TestForwardingParity.

No NetCDF override touches any cell method, so nothing in
src/pyramids/netcdf/ needed updating. Full main test suite passes;
only the two pre-existing Windows file-lock flakes in
test_dataset_collection.py remain failing.
