+++ src/pyramids/dataset/ops/__init__.py
@@ -2,7 +2,6 @@
 
 from pyramids.dataset.ops.analysis import Analysis
 from pyramids.dataset.ops.band_metadata import BandMetadata
-from pyramids.dataset.ops.cell import Cell
 from pyramids.dataset.ops.cog import COGMixin
 from pyramids.dataset.ops.io import IO
 from pyramids.dataset.ops.spatial import Spatial
@@ -12,7 +11,6 @@ __all__ = [
     "Analysis",
     "BandMetadata",
     "COGMixin",
-    "Cell",
     "IO",
     "Spatial",
     "Vectorize",
